# docs/swagger.yaml
swagger: "2.0"
info:
  description: "Documentation of PES-Decobarri"
  version: "1.0.0"
  title: "DecoBarri"
host: project-pes.herokuapp.com
tags:
- name: User
  description: ""
- name: Project
  description: ""
schemes:
- "https"
paths:
  /user/findAll:
    get:
      summary: " "
      description: " "
      tags:
      - User
      produces:
      - "application/json"
      responses:
        200:
          description: "Successful operation"
        500:
          description: "Internal Server Error"
  
  /user/findByID/{username}:
    get:
      tags:
      - User
      parameters:
        - name: username
          in: path
          description: ID de l'usuari
          required: true
          type: string
      responses:
        200:
          description: Success
        500:
          description: Internal Server Error
        404:
          description: User not found.
          
  /user/findByName/{name}:
    get:
      tags:
      - User
      parameters:
        - name: name
          in: path
          description: ID de l'usuari
          required: true
          type: string
      responses:
        200:
          description: Success
        500:
          description: Internal Server Error
    
  /user/add:
    post:
      tags:
      - User
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/CreateUser'
      responses:
        200:
          description: Success
        409:
          description: User alredy registered
        500:
          description: "Internal Server Error"
    
  /user/delete/{username}:
    delete:
      tags:
      - User
      parameters:
        - name: username
          in: path
          description: ID de l'usuari
          required: true
          type: string
      responses:
        200:
          description: User deleted
        404:
          description: User not found
        500:
          description: Internal Server Error
    
  /user/login:
    post:
      tags:
      - User
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Login'
      responses:
        200:
          description: Retorna my Token
        401:
          description: Invalid password
        404:
          description: User not found
        500:
          description: Internal Server Error
    
  /user/edit/{username}:
    put:
      tags: 
        - User
      parameters:
        - name: username
          in: path
          description: ID de l'usuari
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/EditUser'
      responses:
        200:
          description: Success
        401:
          description: Invalid password
        404:
          description: User not found
        500:
          description: Internal Server Error
  
  /user/addProject/{username}:
    put:
      tags: 
        - User
      parameters:
        - name: username
          in: path
          description: ID de l'usuari
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/AddProjectToUser'
      responses:
        200:
          description: Success
        404:
          description: User not found
        500:
          description: Internal Server Error
  
  /project/findAll:
    get:
      tags:
      - Project
      produces:
      - "application/json"
      responses:
        200:
          description: "Successful operation"
        500:
          description: "Internal Server Error"
  
  /project/findByID/{username}:
    get:
      tags:
      - Project
      parameters:
        - name: username
          in: path
          required: true
          type: string
      responses:
        200:
          description: Success
        500:
          description: Internal Server Error
        404:
          description: Project not found.

  /project/findByName/{name}:
    get:
      tags:
      - Project
      parameters:
        - name: name
          in: path
          required: true
          type: string
      responses:
        200:
          description: Success
        500:
          description: Internal Server Error
         
  /project/add:
    post:
      tags:
      - Project
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/StructureProject'
      responses:
        200:
          description: Success. Returns Project id
        500:
          description: "Internal Server Error"

  /project/delete/{username}:
    delete:
      tags:
      - Project
      parameters:
        - name: username
          in: path
          description: ID del projecte
          required: true
          type: string
      responses:
        200:
          description: Project deleted
        404:
          description: Project not found
        500:
          description: Internal Server Error
          
  /project/edit/{username}:
    put:
      tags: 
        - User
      parameters:
        - name: username
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/StructureProject'
      responses:
        200:
          description: Success
        401:
          description: Invalid password
        404:
          description: User not found
        500:
          description: Internal Server Error
  
  /material/findAll:
    get:
      summary: " "
      description: " "
      tags:
      - Material
      produces:
      - "application/json"
      responses:
        200:
          description: "Succes. Returns username of material"
        500:
          description: "Internal Server Error"
  
  /material/add:
    post:
      tags:
      - Material
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/StructureMaterial'
      responses:
        200:
          description: Success. Returns Material id
        500:
          description: "Internal Server Error"

  /material/edit/{username}:
    put:
      tags: 
        - Material
      parameters:
        - name: username
          in: path
          required: true
          type: string
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/StructureMaterial'
      responses:
        200:
          description: Success
        400:
          description: Param required.
        401:
          description: Param not the right type.
        404:
          description: Material not found
        500:
          description: Internal Server Error

definitions:
  CreateUser: 
    type: object
    properties:
      username:
        type: string
      name:
        type: string
      password:
        type: string
      email:
        type: string
        
  EditUser: 
    type: object
    properties:
      name:
        type: string
      old_password:
        type: string
      new_password:
        type: string
      email:
        type: string
        
  Login:
    type: object
    properties:
      username:
        type: string
      password:
        type: string

  AddProjectToUser:
    type: object
    properties:
      project_id:
        type: string

  StructureProject: 
    type: object
    properties:
      name:
        type: string
      theme:
        type: string
      description:
        type: string
      city:
        type: string
      address:
        type: string

  StructureMaterial: 
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      urgent:
        type: boolean
      quantity:
        type: integer
      address:
        type: string